Update coadd unit test for new algMetadata behaivor
Each time the run method of a measurement task is executed, algorithm
metadata is appended to the algorithm metadata object. Previously the run
method of a measurement task was run multiple times for each dataID (due
to limitations in applying aperture corrections). This lead to a behavior
where every dataID had multiple entries for algorithm metadata. However,
now each dataID now only adds one entry of metadata, and multiple entries
will only be added by successive runs of multiple dataIDS. Because the
number of values we be determined by the number of times the run method
is executed, updates are necessary to handle varying numbers.